requirements.txt
argon2-cffi==19.2.0
astroid==2.3.3
autopep8==1.5
BTrees==4.7.1
cffi==1.14.0
falcon==2.0.0
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
persistent==4.6.4
pycodestyle==2.5.0
pycparser==2.20
pylint==2.4.4
PyYAML==5.3
six==1.14.0
transaction==3.0.0
wrapt==1.11.2
zc.lockfile==2.0
ZConfig==3.5.0
ZODB==5.5.1
zodbpickle==2.0.0
zope.interface==5.0.1
